+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - PR #2437 ORC reader: Add 'use_np_dtypes' option
 - PR #2382 Add CategoricalAccessor add, remove, rename, and ordering methods
 - PR #2449 Java column vector: added support for getting byte count of strings in a ColumnVector 
+- PR #2431 Fix local build generated file ownerships
 
 ## Improvements